Add support for AC voltages
Currently only support for sin/cos/anything in between is supported,
but I plan on adding support for square waves, sawtooth, etc. in the
future.

The code is kind of bad and I want to refactor it. I treat constant
voltage sources and non-constants ones and different types, but this
is stupid and adds a lot of unneeded complexity. So, Ill change it to
one component type in the future.