Remove the need for a seperate function component
Makes the flow better, now both constant and variable voltage sources
use the same logic. This might make it a bit slower, cause we will
have to evaluate a function that will always give the same result per
DAE step, but that is a sacrifice I am willing to make.