Add support for fourier transforms
We can now call fourier_transform to plot the DFT of a varible. At the
moment we take in two arguments, how ever the first one does nothing,
this is because I plan on making the output of the DFT a new varible
that can be used/manipulated later, but this might not work out.