Change the low pass 100Hz filter example
I added a AC voltage source in series and added a fourier transform
for the input and output. Now this illistrates what is being done in a
more elegent way. We can see that the out put clearly has less high
frequency data, which can also be seen in the fourier magnitude plot.