Add suppport for calculate function
The calculate function takes in two major arguments, the opperation
and the output name. Depending on the function it will then require
more arguments after. At the moment only the VOLTAGE opperation is
full supported, but the skeleton for CURRENT also exists.
In doing so I have had to majorly change how the data is stored and
manipulated though out the process. I created a data token, which
stores the data, this allows me to point to the data token and say
this is the data that we are going to do the calculation on. Further
more this should not require redundant data and will allow for more
advance opperations, for example, we could take the fourier transform
of two voltages and then add them together and plot it.