Add support for current and voltage calculation for inductors
Added an integrate with time function, as we have to solve an integral
to get the current through the inductor. Although I just realised when
writing this we have already solved for the current during the
differential equation solving portion of the program. So this is not
required, but stil works :/