Add support for diodes in the circuit class
Still need to deal with calculating it, which seems like the hard
part.

Ideas for calculating:
It seem that we can use Shockley's equation, this will require
converting A into a function matrix, which should be fine. But the
hard part seems to be when we use Newton-Raphson Method to find the
roots, here we will need to know the derivative of Shockley's
equation. This seems like a hard thing to do well, I will have to
write a function that takes the derivative of any function, if I feel
like doing it properly.
From:
[https://en.wikipedia.org/wiki/Modified_nodal_analysis#Non-smooth_analysis]
It may not be easy to use the the more ideal equation. So Shockley's
equation will have to do.