Fix inductor not working in some cases + examples
I also discovered that we cannot connect a capacitor to a voltage
source directly, this results in a the resulting E matrix being
singular. There needs to be a resistance in between the capacitor and
voltage source for it to be solveable, this makes sense when comparing
to the real world, as there is some small resistance in any voltage source.