Code clean up
There were flycheck errors with std::numbers for some reason, so we
are now using M_PI.

Add more error handling to the file parser. Now if a mistake is made
it should be more obvious where and what we did wrong.

Modifed the examples stop time and time step. The previous ones were
just place holders and are not what should be used, I still need to
change others, but that can be done later.