Add support for operational amplifiers
It is very scuffed at the moment, I need to fix up a lot of the logic
with creating the DAE, but for this specific example it is working.
The example that was added is a simple not inverting amplifier with a
gain of 6. It turns the constant 5V soruce into a constant 30V source.